moves filter to use location
